+++ apps/web/src/components/modals/alertModal.tsx
@@ -12,8 +12,8 @@ import {
 } from "~/components/ui/alert-dialog";
 
 import { Input } from "../ui/input";
-import { useTranslation } from "~/app/i18n/client";
-import type { SupportedLanguages } from "~/app/i18n/settings";
+import { useTranslation } from "~/i18n/client";
+import type { SupportedLanguages } from "~/i18n/settings";
 
 type AlertModalProps = {
   isOpen: boolean;
