+++ apps/web/src/components/navbar.tsx
@@ -14,8 +14,8 @@ import {
   DropdownMenuItem,
   DropdownMenuTrigger,
 } from "./ui/dropdown-menu";
-import type { SupportedLanguages } from "~/app/i18n/settings";
-import { useTranslation } from "~/app/i18n/client";
+import type { SupportedLanguages } from "~/i18n/settings";
+import { useTranslation } from "~/i18n/client";
 
 interface NavbarProps {
   lang: SupportedLanguages;
