+++ apps/web/src/i18n/locales/en/settings.json
