+++ apps/web/src/i18n/locales/en/translation.json
