+++ apps/web/src/i18n/locales/pl/settings.json
