+++ apps/web/src/i18n/settings.ts
